+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/preferences/MamePreferencesController.java
@@ -1,25 +1,16 @@
 package de.mephisto.vpin.ui.preferences;
 
 import de.mephisto.vpin.commons.utils.WidgetFactory;
-import de.mephisto.vpin.restclient.components.ComponentRepresentation;
-import de.mephisto.vpin.restclient.components.ComponentType;
 import de.mephisto.vpin.restclient.mame.MameOptions;
 import de.mephisto.vpin.ui.PreferencesController;
 import de.mephisto.vpin.ui.Studio;
-import de.mephisto.vpin.ui.events.EventManager;
-import de.mephisto.vpin.ui.util.Dialogs;
-import javafx.application.Platform;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
-import javafx.scene.control.Button;
-import javafx.scene.control.ButtonType;
 import javafx.scene.control.CheckBox;
-import javafx.scene.control.Label;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.net.URL;
-import java.util.Optional;
 import java.util.ResourceBundle;
 
 import static de.mephisto.vpin.ui.Studio.client;
@@ -59,57 +50,6 @@ public class MamePreferencesController implements Initializable {
 
   @FXML
   private CheckBox soundMode;
-
-  @FXML
-  private Button setVersionBtn;
-
-  @FXML
-  private Button installBtn;
-
-  @FXML
-  private Button refreshBtn;
-
-  @FXML
-  private Label installedVersion;
-
-  @FXML
-  private Label latestVersion;
-
-  private ComponentRepresentation component;
-
-  @FXML
-  private void onInstall() {
-    Dialogs.openComponentUpdateDialog(ComponentType.vpinmame, "Installation of \"VPin MAME " + this.latestVersion.getText() + "\"");
-  }
-
-  @FXML
-  private void onVersionRefresh() {
-    refreshBtn.setDisable(true);
-
-    new Thread(() -> {
-      client.getComponentService().clearCache();
-
-      Platform.runLater(() -> {
-        EventManager.getInstance().notify3rdPartyVersionUpdate();
-        this.refreshUpdate();
-      });
-      refreshBtn.setDisable(false);
-    }).start();
-  }
-
-  @FXML
-  private void onVersionSet() {
-    Optional<ButtonType> result = WidgetFactory.showConfirmation(Studio.stage, "Set Version", "Apply \"" + component.getLatestReleaseVersion() + "\" as the current version of VPin MAME?", null, "Apply");
-    if (result.isPresent() && result.get().equals(ButtonType.OK)) {
-      try {
-        client.getComponentService().setVersion(ComponentType.vpinmame, component.getLatestReleaseVersion());
-        EventManager.getInstance().notify3rdPartyVersionUpdate();
-      } catch (Exception e) {
-        WidgetFactory.showAlert(Studio.stage, "Error", "Failed to apply version: " + e.getMessage());
-      }
-      refreshUpdate();
-    }
-  }
 
   private void saveOptions() {
     MameOptions options = new MameOptions();
@@ -165,15 +105,5 @@ public class MamePreferencesController implements Initializable {
     colorizeDmd.selectedProperty().addListener((observable, oldValue, newValue) -> saveOptions());
     soundMode.setSelected(options.isColorizeDmd());
     soundMode.selectedProperty().addListener((observable, oldValue, newValue) -> saveOptions());
-
-    refreshUpdate();
-  }
-
-  private void refreshUpdate() {
-    component = client.getComponentService().getComponent(ComponentType.vpinmame);
-    if (component != null) {
-      installedVersion.setText(component.getInstalledVersion() != null ? component.getInstalledVersion() : "?");
-      latestVersion.setText(component.getLatestReleaseVersion() != null ? component.getLatestReleaseVersion() : "?");
-    }
   }
 }