+++ RELEASE_NOTES.md
@@ -3,6 +3,12 @@
 
 ### Changes
 
+**Preferences**
+
+- Combined "Service Info" and "Service Options" to "VPin Studio Service" entry.
+- Renamed section to "VPin System Preferences"
+- Add entry "System Manager"
+
 **Table Management: DMD**
 
 - Added bulk operation to execute the "Auto-Fill" for all tables.